# perf: disable vite file watcher