We don't use Vite watch functionality in run mode and setting up file watcher takes considerable time:
![image](https://user-images.githubusercontent.com/2339406/147922397-f54c3071-96ad-4a51-b4bd-e679666a4e71.png)


Before:

| name   | time  |
|---|---|
|  vitest:vue | '4.147s ± 12.89%' |
| jest:vue | '4.629s ± 1.96%'  |

After:
| name | time |
|---|---|
| vitest:vue | '3.438s ± 1.84%' |
| jest:vue | '4.658s ± 3.04%' |