@dominikg 2022-01-03
are you sure? https://github.com/vitest-dev/vitest/blob/16ccf14d6ba311b9ced1cb57973b55a4e7d51ad7/packages/vitest/src/node/config.ts#L48

I think this could be added as an optimization if config.watch is false