@dominikg 2022-01-03
even more useful could be an option to completely disable vite's own watcher in that case. It seems counter-intuitive to have a vite dev server that isn't watching for file changes but for vitests usecase it could make sense.