@Demivan 2022-01-03
> I think this could be added as an optimization if config.watch is false

@dominikg You are right. Strange it was working for me.

I have checked Vite source and disabling file watcher is a very big change. Just setting ignored files to `'**'` negates its performance impact on Vitest.